requirements.txt
# SGR Research Agent - Production Dependencies
# Основные зависимости для работы системы Schema-Guided Reasoning

# OpenAI API для работы с языковыми моделями
openai>=1.0.0

# Tavily для веб-поиска с фокусом на достоверность
tavily-python>=0.3.0

# Pydantic для валидации схем и структурированных ответов
pydantic>=2.0.0

# Rich для красивого консольного интерфейса
rich>=13.0.0

# PyYAML для обработки конфигурационных файлов
PyYAML>=6.0

# Trafilatura для извлечения текста с веб-страниц (с встроенным readability)
trafilatura>=1.6.0

# YouTube Transcript API для получения субтитров YouTube видео
youtube-transcript-api>=1.0.0

rapidfuzz>=3.9
outlines>=0.0.46
# Дополнительные зависимости для разработки (опционально)
# pytest>=7.0.0          # Для тестирования
# black>=23.0.0          # Для форматирования кода
# flake8>=6.0.0          # Для линтинга
# mypy>=1.0.0            # Для проверки типов